Move attributes listener to own namespace
